move resolve_base_url into network.py; drop connection module
